refactor(versioning): split baseline.py into a package
``superset.versioning.baseline`` was a 633-LOC single module that read
top-down in stepdown order from the public entry point through the
leaf helpers. Convert it to a 7-module package along the existing
banner-comment dividers:

* ``baseline/shadow.py`` (~75 LOC) — low-level
  :func:`_insert_baseline_shadow_row` writer +
  :data:`CONTINUUM_BOOKKEEPING_COLUMNS` constant.
* ``baseline/dirty.py`` (~150 LOC) — promote-parent-to-dirty
  machinery (:func:`_force_parent_dirty_on_child_change`,
  :func:`_pin_audit_columns`).
* ``baseline/collection.py`` (~130 LOC) — discovery: which parents
  need a baseline (:func:`_collect_parents_to_baseline`,
  :func:`_child_to_parent_registry`, :func:`_version_table_for`,
  :func:`_shadow_row_count`) plus ``VERSIONED_MODELS``.
* ``baseline/insertion.py`` (~135 LOC) — top-level
  :func:`_insert_baseline_and_children` glue +
  :func:`_insert_baseline_row` for parent shadow rows +
  :func:`_baseline_children_for_parent` dispatch.
* ``baseline/children.py`` (~185 LOC) — per-entity child handlers
  (:func:`_baseline_dataset_children` /
  :func:`_baseline_dashboard_children`),
  :data:`_CHILD_BASELINE_HANDLERS`, and the leaf-level row writers
  (:func:`_insert_child_baseline_rows`,
  :func:`_baseline_attached_slices`,
  :func:`_insert_synthetic_slice_baseline`).
* ``baseline/listener.py`` (~75 LOC) — public
  :func:`register_baseline_listener` that wires the ``before_flush``
  event on ``db.session``.
* ``baseline/__init__.py`` (~70 LOC) — re-exports for backward
  compat.

The six external import sites (initialization wiring, the changes
package's ``CONTINUUM_BOOKKEEPING_COLUMNS`` consumer, ``factory.py``'s
``_child_to_parent_registry`` consumer, ``queries.py``,
``test_pin_audit_columns``) keep working unchanged via the
``__init__.py`` re-exports.

Listener registration order is preserved — the single
``event.listens_for(db.session, "before_flush", insert=True)``
declaration in :func:`register_baseline_listener` is byte-for-byte
the same; only the function's file location changed.

Smoke-tested inside the container: all 6 public symbols import
cleanly via ``from superset.versioning.baseline import …``.

Surfaced by superset-committer-review v2.
